+++ tests/AdventOfCode.Tests/Puzzles/Y2019/Day02Tests.cs
@@ -23,15 +23,15 @@ namespace MartinCostello.AdventOfCode.Puzzles.Y2019
         }
 
         [Theory]
-        [InlineData("1,9,10,3,2,3,11,0,99,30,40,50", new[] { 3500, 9, 10, 70, 2, 3, 11, 0, 99, 30, 40, 50 })]
-        [InlineData("1,0,0,0,99", new[] { 2, 0, 0, 0, 99 })]
-        [InlineData("2,3,0,3,99", new[] { 2, 3, 0, 6, 99 })]
-        [InlineData("2,4,4,5,99,0", new[] { 2, 4, 4, 5, 99, 9801 })]
-        [InlineData("1,1,1,4,99,5,6,0,99", new[] { 30, 1, 1, 4, 2, 5, 6, 0, 99 })]
-        public void Y2019_Day02_RunProgram_Returns_Correct_State(string program, int[] expected)
+        [InlineData("1,9,10,3,2,3,11,0,99,30,40,50", new[] { 3500, 9, 10, 70, 2, 3, 11, 0, 99, 30, 40, 50L })]
+        [InlineData("1,0,0,0,99", new[] { 2, 0, 0, 0, 99L })]
+        [InlineData("2,3,0,3,99", new[] { 2, 3, 0, 6, 99L })]
+        [InlineData("2,4,4,5,99,0", new[] { 2, 4, 4, 5, 99, 9801L })]
+        [InlineData("1,1,1,4,99,5,6,0,99", new[] { 30, 1, 1, 4, 2, 5, 6, 0, 99L })]
+        public void Y2019_Day02_RunProgram_Returns_Correct_State(string program, long[] expected)
         {
             // Act
-            IReadOnlyList<int> actual = Day02.RunProgram(program);
+            IReadOnlyList<long> actual = Day02.RunProgram(program);
 
             // Assert
             actual.ShouldBe(expected);
